requirements.txt
faststream[rabbit]
pydantic-settings
aiohttp
aio-pika
gunicorn==21.2.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
black==25.1.0
click==8.2.0
fastapi==0.115.2
GeoAlchemy2==0.15.2
greenlet==3.2.2
h11==0.16.0
idna==3.10
loguru==0.7.3
python-logging-loki==0.3.1
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.13.2
uvicorn==0.32.0